more web and angular

+++ Imperatur_v2/order/OrderQueue.cs
@@ -173,32 +173,36 @@ namespace Imperatur_v2.order
 
 
             ToRemove.Clear();
-            List<IOrder> ToAdd = new List<IOrder>();
+            
             bool bReturn = false;
-            foreach (IOrder oI in m_oOrders.Where(x => x != null).OrderBy(x => x.ValidToDate))
+            if (m_oOrders.Count() > 0 && m_oOrders.Where(x => x != null).Count() > 0)
             {
-                IOrder StopLoss;
-                try
+                List<IOrder> ToAdd = new List<IOrder>();
+                foreach (IOrder oI in m_oOrders.Where(x => x != null).OrderBy(x => x.ValidToDate))
                 {
-                    if (oI.ExecuteOrder(m_oAccountHandler, m_oTradeHandler, out StopLoss))
+                    IOrder StopLoss;
+                    try
                     {
-                        bReturn = true;
-                        if (!oI.Equals(StopLoss))
+                        if (oI.ExecuteOrder(m_oAccountHandler, m_oTradeHandler, out StopLoss))
                         {
-                            ToAdd.Add(StopLoss);
+                            bReturn = true;
+                            if (!oI.Equals(StopLoss))
+                            {
+                                ToAdd.Add(StopLoss);
+                            }
+                            ToRemove.Add(oI);
                         }
-                        ToRemove.Add(oI);
                     }
+                    catch (Exception ex)
+                    {
+                        ImperaturGlobal.GetLog().Error("Exception in OrderQueue.EvaluateOrdersInQueue", ex);
+                    }
+
                 }
-                catch(Exception ex)
-                {
-                    int gg = 0;
-                }
-                  
+                m_oOrders.RemoveRange(ToRemove);
+                RemoveFilesFromStorage(ToRemove);
+                m_oOrders.AddRange(ToAdd);
             }
-            m_oOrders.RemoveRange(ToRemove);
-            RemoveFilesFromStorage(ToRemove);
-            m_oOrders.AddRange(ToAdd);
             return bReturn;
         }
